Storybook 7 migration changes
