+++ packages/storybook/package.json
@@ -15,7 +15,8 @@
     "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ../../.prettierrc --ignore-path=../../.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
     "lint": "node ./scripts/lint.mjs",
     "lint:fix": "rushx lint -- --fix",
-    "lint:quiet": "rushx lint -- --quiet"
+    "lint:quiet": "rushx lint -- --quiet",
+    "prepare": "husky install"
   },
   "dependencies": {
     "@azure/communication-calling": "1.14.1-beta.1 || >=1.14.1",
@@ -23,20 +24,20 @@
     "@azure/communication-common": "^2.2.0",
     "@azure/communication-identity": "^1.2.0",
     "@azure/communication-react": "1.6.1-beta.0",
+    "@fluentui/react": "~8.98.3",
     "@fluentui/react-file-type-icons": "8.8.13",
     "@fluentui/react-hooks": "~8.6.12",
     "@fluentui/react-icons": "~2.0.194",
     "@fluentui/react-northstar": "0.61.0",
-    "@fluentui/react": "~8.98.3",
     "@fluentui/theme-samples": "8.1.5",
     "@internal/fake-backends": "1.6.1-beta.0",
     "@microsoft/applicationinsights-react-js": "~3.0.5",
     "@microsoft/applicationinsights-web": "~2.6.2",
+    "browserslist": "~4.20.4",
     "copy-to-clipboard": "~3.3.1",
     "resize-observer-polyfill": "~1.5.1",
     "tslib": "~2.5.0",
-    "uuid": "^8.1.0",
-    "browserslist": "~4.20.4"
+    "uuid": "^8.1.0"
   },
   "devDependencies": {
     "@babel/core": "~7.16.0",
@@ -44,30 +45,29 @@
     "@mdx-js/react": "^1.6.22",
     "@microsoft/api-documenter": "~7.12.11",
     "@microsoft/api-extractor": "~7.18.0",
-    "@storybook/addons": "~6.5.14",
-    "@storybook/api": "~6.5.14",
-    "@storybook/core-events": "~6.5.14",
-    "@storybook/components": "~6.5.14",
-    "@storybook/addon-actions": "~6.5.14",
-    "@storybook/addon-controls": "~6.5.14",
-    "@storybook/addon-docs": "~6.5.14",
-    "@storybook/addon-essentials": "~6.5.14",
-    "@storybook/addon-links": "~6.5.14",
-    "@storybook/addon-storyshots": "~6.5.14",
-    "@storybook/builder-webpack5": "6.5.14",
-    "@storybook/manager-webpack5": "6.5.14",
-    "@storybook/node-logger": "~6.5.14",
-    "@storybook/react": "~6.5.14",
+    "@storybook/addon-actions": "~7.0.23",
+    "@storybook/addon-controls": "~7.0.23",
+    "@storybook/addon-docs": "~7.0.23",
+    "@storybook/addon-essentials": "~7.0.23",
+    "@storybook/addon-links": "~7.0.23",
+    "@storybook/addon-storyshots": "~7.0.23",
+    "@storybook/addons": "~7.0.23",
+    "@storybook/api": "~7.0.23",
+    "@storybook/components": "~7.0.23",
+    "@storybook/core-events": "~7.0.23",
+    "@storybook/node-logger": "~7.0.23",
+    "@storybook/react": "~7.0.23",
+    "@storybook/react-webpack5": "~7.0.23",
     "@storybook/storybook-deployer": "~2.8.16",
-    "@storybook/theming": "~6.5.14",
+    "@storybook/theming": "~7.0.23",
     "@testing-library/jest-dom": "^5.11.4",
     "@types/classnames": "^2.2.11",
     "@types/jest": "~26.0.22",
     "@types/json-stringify-safe": "^5.0.0",
     "@types/node": "^16.11.7",
+    "@types/react": "18.2.0",
     "@types/react-dom": "18.2.1",
     "@types/react-linkify": "^1.0.0",
-    "@types/react": "18.2.0",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
@@ -77,29 +77,29 @@
     "concurrently": "^5.3.0",
     "copyfiles": "^2.4.1",
     "core-js": "^3.8.3",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-header": "^3.1.0",
     "eslint-plugin-import": "~2.22.1",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-mdx": "1.16.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react-hooks": "^4.1.2",
     "eslint-plugin-react": "^7.18.3",
-    "eslint": "^7.7.0",
+    "eslint-plugin-react-hooks": "^4.1.2",
+    "gulp": "~4.0.2",
     "husky": "^4.3.0",
+    "jest": "26.6.0",
     "jest-fetch-mock": "^3.0.3",
     "jest-junit": "~13.0.0",
-    "jest": "26.6.0",
     "json-stringify-safe": "^5.0.1",
     "nan": "^2.14.1",
     "node-forge": ">=1.0.0",
-    "preact": "^10.5.12",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
     "raw-loader": "~4.0.2",
+    "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-is": "~17.0.1",
-    "react": "18.2.0",
     "regenerator-runtime": "^0.13.7",
     "require-from-string": "^2.0.2",
     "scheduler": "^0.20.0",
